refactor(tests): split bridge contract-runner validations (#1401)
